support a configurable libvirt injection partition
This is useful if all guest images have the same structure,
and the root partition is not the first partition.

This is also handy to enable inspection in libguestfs,
which can handle disparate and complicated image layouts.

In future we may change to a StrOpt to support
searching by partition label.

Change-Id: Ie94d61bec8fe4b41d6d2d6d3efa9a4364cf027fe

Conflicts:

	nova/virt/libvirt/connection.py