libvirt: Fix root disk leak in live mig
This patch makes sure that _create_images_and_backing method of the
libvirt driver (called in several places, but most problematic one is
the call in the pre_live_migration method) creates all the files the
instance needs that are not present.

Prioir to this patch - the method would only attempt to download the
image, and if it did so with the path of the ephemeral drives, it could
expose the image to other users as an ephemeral devices. See the related
bug for more detaiis.

After this patch - we properly distinguish between image, ephemeral and
swap files, and make sure that the imagebackend does the correct thing.

Closes-bug: #1251590

Co-authored-by: Loganathan Parthipan <parthipan@hp.com>

This patch also includes part of commit
65386c91910ee03d947c2b8bcc226a53c30e060a, not cherry-picked as a whole
due to the fact that it is a trivial change, and to avoud the
proliferation of patches needed to fix this bug.

(cherry picked from commit c69a619668b5f44e94a8fe1a23f3d887ba2834d7)

Conflicts:
	nova/tests/test_libvirt.py
	nova/virt/libvirt/driver.py

Change-Id: I78aa2f4243899db4f4941e77014a7e18e27fc63e